vigenere encipher and decipher program
I have also included a test text file for your enjoyment.  the passcode is secret.
currently the vigenere algorithm is mostly implemented in main().  if it is to be called as a modular function, it will take a little re-working, maybe needs to be rewritten using a class to allow data persistence of the codephrase counter between calls to function.

enjoy!  what I'm thinking of doing as a puzzle if anyone other than me would want to try, is to encipher a longer text in Vigenere code and have you try to decipher it without the codephrase.  that would be a true challenge.  see The Code Book for how to crack this.